fix(collection): updated collection classes

+++ src/main/kotlin/org/example/kotlincollectionmanager/collection/items/Coordinates.kt
@@ -1,8 +1,8 @@
 package org.example.kotlincollectionmanager.collection.items
 
-class Coordinates(private var x: Long, private var y: Long) {
+class Coordinates(private var x: Long, private var y: Float) {
     fun getX(): Long = x
-    fun getY(): Long = y
+    fun getY(): Float = y
 
     override fun toString(): String {
         return "Coordinates(x=$x, y=$y)"
